Adds more CQCode Support, fix CQCode escape, code from iTXTech/mirai-native
