+++ src/main/kotlin/tech/mihoyo/mirai/data/common/MessageDTO.kt
@@ -17,6 +17,7 @@ import net.mamoe.mirai.contact.Contact
 import net.mamoe.mirai.contact.Group
 import net.mamoe.mirai.message.*
 import net.mamoe.mirai.message.data.*
+import net.mamoe.mirai.utils.MiraiExperimentalAPI
 import net.mamoe.mirai.utils.currentTimeMillis
 import java.net.URL
 
@@ -63,12 +64,14 @@ data class CQPrivateMessagePacketDTO(
 
 
 // Message toCQString
+@MiraiExperimentalAPI
 suspend fun Message.toCQString(): String {
     return when (this) {
         is PlainText -> content
         is At -> "[CQ:at,qq=$target]"
         is Face -> "[CQ:face,id=$id]"
-        is PokeMessage -> "[CQ:shake]"
+        is VipFace -> "[CQ:vipface,id=${kind.id},name=${kind.name},count=${count}]"
+        is PokeMessage -> "[CQ:poke,id=${id},type=${type},name=${name}]"
         is AtAll -> "[CQ:at,qq=all]"
         is Image -> {
             "[CQ:image,file=$imageId,url=${queryUrl()}]"
@@ -76,6 +79,7 @@ suspend fun Message.toCQString(): String {
         is RichMessage -> "[CQ:rich,data=${content}]"
         is MessageSource -> ""
         is QuoteReply -> ""
+        is Voice -> "[CQ:voice,url=${url},md5=${md5},file=${fileName}]"
         else -> "此处消息的转义尚未被插件支持"
     }
 }
